refactor: move JS interop code to `jstypes` subpackage
