+++ appinventor/appengine/src/com/google/appinventor/client/utils/jstypes/Blob.java
@@ -1,4 +1,4 @@
-package com.google.appinventor.client.utils;
+package com.google.appinventor.client.utils.jstypes;
 
 import jsinterop.annotations.JsConstructor;
 import jsinterop.annotations.JsPackage;
