+++ appinventor/appengine/src/com/google/appinventor/client/utils/jstypes/URL.java
@@ -1,11 +1,11 @@
-package com.google.appinventor.client.utils;
+package com.google.appinventor.client.utils.jstypes;
 
 import jsinterop.annotations.JsMethod;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 
-@JsType(isNative = true, namespace = JsPackage.GLOBAL, name = "URL")
-public class JsURL {
+@JsType(isNative = true, namespace = JsPackage.GLOBAL)
+public class URL {
   @JsMethod
   public static native String createObjectURL(Object object);
 }
